Set `EMSDK_PYTHON=../python3` to match what `emsdk_env.sh` does
We add the path containing the python binaries from the packages to
`PATH`, but that contains `python3`, and not `python`. When we run
`emcc --version` on a machine without a system python, it fails with

    `unable to find python in $PATH`

This was not caught with our regular testing, possibly because a system
`python` was available on those machines.

`emsdk_env.sh` sets `EMSDK_PYTHON = /Users/radical/dev/r3/src/mono/wasm/emsdk/python/3.7.4-2_64bit/bin/python3`

.. so, this commit adds the same for running `emcc`, and sets that to
`python3` from the package.

This was tested on macOS.