Fix path issues with wixpacks (#54)
Temporary fix for the blob path issues for the wixpacks. Previously the missing _UploadPathRoot was causing paths like:
`/wixpack//foofile.6.0.0-preview.1234.wixpack.zip`
This causes problems in gather drop due to Path.Combine behavior (the rooted path wins and the file is placed in the root of the directory structure).
This is not a full solution, because when stable builds happen, these paths must be unique. At that point, the `foofile.6.0.0-preview.1234.wixpack.zip` will end up foofile.6.0.0.wixpack.zip, which
will be non-unique in blob storage.
Fixing this requires a bit of messing about to get full version info in the publishing targets, so deferring that for now and unblocking signing.